+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Derive `Clone` on `ModelInjector` even if the model doesn't implement clone
   ([#165]).
+- Introduce an `EventInjector` bound to a single model input ([#167]).
+- Add a possibility to inject queries via the global `Injector`.
 
 # 1.0.0 (2026-02-03)